Move numba core shape test next to the numba rewrite tests
